+++ Tensile/Source/client/include/ClientProblemFactory.hpp
@@ -70,18 +70,18 @@ namespace Tensile
             ContractionProblem::BatchIndices m_batchIndices;
             ContractionProblem::BoundIndices m_boundIndices;
 
-            DataType        m_aType;
-            DataType        m_bType;
-            DataType        m_cType;
-            DataType        m_dType;
-            DataType        m_alphaType;
-            DataType        m_betaType;
-            bool            m_stridedBatched;
-            bool            m_highPrecisionAccumulate;
-            bool            m_deterministicMode;
-            ArithmeticUnit  m_arithmeticUnit;
-            BenchmarkMetric m_benchmarkMetric;
-            KernelLanguage  m_kernelLanguage;
+            DataType          m_aType;
+            DataType          m_bType;
+            DataType          m_cType;
+            DataType          m_dType;
+            DataType          m_alphaType;
+            DataType          m_betaType;
+            bool              m_stridedBatched;
+            bool              m_highPrecisionAccumulate;
+            bool              m_deterministicMode;
+            ArithmeticUnit    m_arithmeticUnit;
+            KernelLanguage    m_kernelLanguage;
+            PerformanceMetric m_performanceMetric;
 
             std::vector<std::vector<size_t>> m_problemSizes;
             std::vector<std::vector<size_t>> m_aStrides;