+++ Tensile/Source/lib/include/Tensile/ContractionProblem.hpp
@@ -25,8 +25,8 @@
 #pragma once
 
 #include <Tensile/ArithmeticUnitTypes.hpp>
-#include <Tensile/BenchmarkMetricTypes.hpp>
 #include <Tensile/KernelLanguageTypes.hpp>
+#include <Tensile/PerformanceMetricTypes.hpp>
 #include <Tensile/Tensile.hpp>
 
 #include <Tensile/ContractionProblem_fwd.hpp>
@@ -544,14 +544,14 @@ namespace Tensile
             return m_kernelLanguage;
         }
 
-        void setBenchmarkMetric(BenchmarkMetric value)
+        void setPerformanceMetric(PerformanceMetric value)
         {
-            m_benchmarkMetric = value;
+            m_performanceMetric = value;
         }
 
-        BenchmarkMetric benchmarkMetric() const
+        PerformanceMetric performanceMetric() const
         {
-            return m_benchmarkMetric;
+            return m_performanceMetric;
         }
 
         void setDeterministicMode(bool value)
@@ -736,15 +736,15 @@ namespace Tensile
         std::string m_sumNames;
         std::string m_operationIdentifier;
 
-        bool            m_transA;
-        bool            m_transB;
-        bool            m_stridedBatched          = true;
-        bool            m_highPrecisionAccumulate = false;
-        bool            m_deterministicMode       = false;
-        bool            m_eligibleForPK           = true;
-        ArithmeticUnit  m_arithmeticUnit          = ArithmeticUnit::Any;
-        BenchmarkMetric m_benchmarkMetric         = BenchmarkMetric::Overall;
-        KernelLanguage  m_kernelLanguage          = KernelLanguage::Any;
+        bool              m_transA;
+        bool              m_transB;
+        bool              m_stridedBatched          = true;
+        bool              m_highPrecisionAccumulate = false;
+        bool              m_deterministicMode       = false;
+        bool              m_eligibleForPK           = true;
+        ArithmeticUnit    m_arithmeticUnit          = ArithmeticUnit::Any;
+        KernelLanguage    m_kernelLanguage          = KernelLanguage::Any;
+        PerformanceMetric m_performanceMetric       = PerformanceMetric::Overall;
 
         DataType m_alphaType = DataType::None; // if not assigned, will follow d-type
         DataType m_betaType  = DataType::None; // for bwd-compatible