+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [(Unreleased) Tensile 4.27.0 for ROCm 4.2.0]
 ### Added
+- Benchmarking and library support for CU efficiency vs. overall speed
 - support general batch GEMM
 - Support offset for each input/output buffer in Tensile
 - support support ldc != ldd for all GEMM kernel